# frontend/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
    "/api/v1/admin/users/{id}/role" : {
      "put" : {
        "tags" : [ "user-admin-controller" ],
        "operationId" : "changeRole",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RoleChange"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AppUser"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/active" : {
      "put" : {
        "tags" : [ "user-admin-controller" ],
        "operationId" : "changeActive",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ActiveChange"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AppUser"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/impact" : {
      "post" : {
        "tags" : [ "impact-controller" ],
        "operationId" : "analyze",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImpactRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImpactResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impact/export" : {
      "post" : {
        "tags" : [ "impact-controller" ],
        "operationId" : "export",
        "parameters" : [ {
          "name" : "format",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImpactRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Credentials"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Me"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/change-password" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "changePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordChange"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Me"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users" : {
      "get" : {
        "tags" : [ "user-admin-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageAppUser"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "user-admin-controller" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Registration"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AppUser"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/csrf" : {
      "get" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "csrf",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/CsrfToken"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CsrfView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/ldap/users" : {
      "get" : {
        "tags" : [ "user-admin-controller" ],
        "operationId" : "searchDirectory",
        "parameters" : [ {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/DirectoryUserView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/audit" : {
      "get" : {
        "tags" : [ "audit-controller" ],
        "operationId" : "list_6",
        "parameters" : [ {
          "name" : "actor",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "action",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageAuditEntry"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "RoleChange" : {
        "type" : "object",
        "properties" : {
          "role" : {
            "type" : "string",
            "enum" : [ "ADMIN", "USER" ]
          }
        }
      },
      "AppUser" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string",
            "enum" : [ "LOCAL", "LDAP" ]
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "ADMIN", "USER" ]
          },
          "roleGrantedBy" : {
            "type" : "string"
          },
          "roleGrantedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastLoginAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
      "ActiveChange" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          }
        }
      },
...
      "ImpactRequest" : {
        "type" : "object",
        "properties" : {
          "symbolIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "changeType" : {
            "type" : "string",
            "enum" : [ "SIGNATURE", "BEHAVIOR" ]
          },
          "depth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "confidences" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "EXACT", "RECOVERED", "NAME_ONLY" ]
            },
            "uniqueItems" : true
          },
          "includeDispatch" : {
            "type" : "boolean"
          }
        }
      },
      "EntryPoint" : {
        "type" : "object",
        "properties" : {
          "symbolId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "key" : {
            "type" : "string"
          },
          "display" : {
            "type" : "string"
          },
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "modulePath" : {
            "type" : "string"
          },
          "label" : {
            "type" : "string"
          },
          "annotation" : {
            "type" : "string"
          },
          "http" : {
            "type" : "boolean"
          },
          "httpMethod" : {
            "type" : "string"
          },
          "httpPath" : {
            "type" : "string"
          }
        }
      },
      "ImpactEdge" : {
        "type" : "object",
        "properties" : {
          "fromSymbolId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "toSymbolId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "CALL", "INSTANTIATION", "METHOD_REF", "TYPE_REF", "EXTENDS", "IMPLEMENTS", "OVERRIDES", "FIELD_READ", "FIELD_WRITE", "ANNOTATION" ]
          },
          "confidence" : {
            "type" : "string",
            "enum" : [ "EXACT", "RECOVERED", "NAME_ONLY" ]
          },
          "level" : {
            "type" : "integer",
            "format" : "int32"
          },
          "viaDispatch" : {
            "type" : "boolean"
          },
          "moduleId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "modulePath" : {
            "type" : "string"
          },
          "filePath" : {
            "type" : "string"
          },
          "line" : {
            "type" : "integer",
            "format" : "int32"
          },
          "column" : {
            "type" : "integer",
            "format" : "int32"
          },
          "snippet" : {
            "type" : "string"
          }
        }
      },
      "ImpactNode" : {
        "type" : "object",
        "properties" : {
          "symbolId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "key" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "CLASS", "INTERFACE", "ENUM", "RECORD", "ANNOTATION_TYPE", "METHOD", "CONSTRUCTOR", "FIELD" ]
          },
          "display" : {
            "type" : "string"
          },
          "level" : {
            "type" : "integer",
            "format" : "int32"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "SEED", "AFFECTED", "DISPATCH", "TWIN" ]
          },
          "confidence" : {
            "type" : "string",
            "enum" : [ "EXACT", "RECOVERED", "NAME_ONLY" ]
          },
          "nameOnly" : {
            "type" : "boolean"
          }
        }
      },
      "ImpactResult" : {
        "type" : "object",
        "properties" : {
          "summary" : {
            "$ref" : "#/components/schemas/ImpactSummary"
          },
          "nodes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImpactNode"
            }
          },
          "edges" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ImpactEdge"
            }
          },
          "entryPoints" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EntryPoint"
            }
          },
          "staleness" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RepoState"
            }
          },
          "versionWarnings" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VersionWarning"
            }
          },
          "truncated" : {
            "type" : "boolean"
          }
        }
      },
      "ImpactSummary" : {
        "type" : "object",
        "properties" : {
          "repositories" : {
            "type" : "integer",
            "format" : "int32"
          },
          "modules" : {
            "type" : "integer",
            "format" : "int32"
          },
          "classes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "methods" : {
            "type" : "integer",
            "format" : "int32"
          },
          "usages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "usagesByLevel" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "usagesByConfidence" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "nodesByConfidence" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "integer",
              "format" : "int32"
            }
          },
          "partialClasspathRepositories" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RepositoryRef"
            }
          }
        }
      },
      "RepoState" : {
        "type" : "object",
        "properties" : {
          "moduleId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "repositoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectKey" : {
            "type" : "string"
          },
          "repository" : {
            "type" : "string"
          },
          "modulePath" : {
            "type" : "string"
          },
          "classpathMode" : {
            "type" : "string"
          },
          "lastIndexedCommit" : {
            "type" : "string"
          },
          "lastIndexedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "RepositoryRef" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "projectKey" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          }
        }
      },
      "VersionWarning" : {
        "type" : "object",
        "properties" : {
          "repository" : {
            "$ref" : "#/components/schemas/RepositoryRef"
          },
          "modulePath" : {
            "type" : "string"
          },
          "dependency" : {
            "type" : "string"
          },
          "usedVersion" : {
            "type" : "string"
          },
          "declaredVersion" : {
            "type" : "string"
          }
        }
      },
      "Credentials" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      },
      "Me" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "ADMIN", "USER" ]
          },
          "source" : {
            "type" : "string",
            "enum" : [ "LOCAL", "LDAP" ]
          },
          "mustChangePassword" : {
            "type" : "boolean"
          }
        },
        "required" : [ "mustChangePassword", "role", "username" ]
      },
      "PasswordChange" : {
        "type" : "object",
        "properties" : {
          "currentPassword" : {
            "type" : "string"
          },
          "newPassword" : {
            "type" : "string"
          }
        }
      },
      "Registration" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "ADMIN", "USER" ]
          }
        }
      },
...
      "CsrfToken" : {
        "type" : "object",
        "properties" : {
          "headerName" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "parameterName" : {
            "type" : "string"
          }
        }
      },
      "CsrfView" : {
        "type" : "object",
        "properties" : {
          "headerName" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        }
      },
      "PageAppUser" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AppUser"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "DirectoryUserView" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "appUserId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "AuditEntry" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "actor" : {
            "type" : "string"
          },
          "action" : {
            "type" : "string"
          },
          "target" : {
            "type" : "string"
          },
          "details" : {
            "type" : "string"
          },
          "at" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageAuditEntry" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AuditEntry"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    }
  }
}